# packages/api-spec/src/crate-spec.openapi.yml
openapi: 3.0.3
info:
  version: "1.0.0"
  title: "Crate Network API"
  description: |
    ## About this spec

    The Crate API exposes several endpoints which allows the Crate macOS, iOS, 
    iPadOS and web clients to interact with Crate's IPFS cluster without the 
    clients needing to run IPFS clients of their own. Crate Network implements 
    the IPFS Pinning API specification at the `/pins` endpoint, as described by 
    the spec found
    [on GitHub](https://github.com/ipfs/pinning-services-api-spec/blob/3196b0b466752a3626b716969703cbd48cb9bcf7/ipfs-pinning-service.yaml). 
    The pinning API accepts a Bearer authorization token retrieved from your
    Crate account. This identifies your user and adds the pinned content to 
    your files on Crate. This version of the API is modified to include 
    metadata, all of which is optional so that Crate continues to conform to 
    the IPFS spec.

    The Filecoin replication service is handled via the pinning API. Instead of 
    using untrusted centralized cloud storage as the backend, Crate safely 
    replicates content onto other decentralized networks. When a request comes 
    to the pinning service to retrieve the content, Crate will prioritize 
    cached content, otherwise falling back to the Filecoin network to retrieve
    your files.
servers:
  - url: https://crate.network/api/v1
paths:
...
  /dir:
    post:
      summary: Create an empty directory.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            description: File path
            example: /ipfs/Qm.../dir1
            type: string
        - name: name
          in: query
          required: true
          schema:
            description: Directory name
            type: string
            example: dir2
      description: |
        Copies the existing structure from the root hash, adds a new named
        folder, and returns the path to it from the new root CID.
      tags:
        - files
      responses:
        "200":
          description: Successful response (path)
          content:
            application/json:
              schema:
                type: string
                description: File path
                example: /ipfs/Qm.../dir1/dir2
  /dir/add:
    post:
      summary: Add a file to a directory.
      tags:
        - files
      parameters:
        - name: path
          in: query
          required: true
          schema:
            description: Directory path
            example: /ipfs/Qm.../dir1/dir2
            type: string
        - name: name
          in: query
          required: true
          schema:
            description: File name
            type: string
            example: file.txt
        - name: cid
          in: query
          required: true
          schema:
            description: CID for file
            type: string
            example: Qm...
      description: |
        Adds an existing file with hash and name specified to a directory.
      responses:
        "200":
          description: Successful response (path)
          content:
            application/json:
              schema:
                type: string
                description: File path
                example: /ipfs/Qm.../dir1/dir2/file.txt
...
components:
...
  securitySchemes:
    accessToken:
      description: |
        An opaque token is required to be sent with each request in the HTTP 
        header:
        - `Authorization: Bearer <access-token>`
        The `access-token` should be generated per device, and the user should 
        have the ability to revoke each token separately.
      type: http
      scheme: bearer
security:
  - accessToken: []